Move query run commands implementation into era-based
