+++ cardano-cli/test/cardano-cli-test/Test/Cli/JSON.hs
@@ -8,7 +8,7 @@ module Test.Cli.JSON
 
 import           Cardano.Api.Shelley
 
-import           Cardano.CLI.Legacy.Run.Query
+import           Cardano.CLI.EraBased.Run.Query
 import           Cardano.CLI.Types.Common
 import           Cardano.CLI.Types.Output (QueryKesPeriodInfoOutput (..), createOpCertIntervalInfo)
